openapi-mingle added models to responses
