+++ test/unit/openapi/atomic-types/openapi.body.object.test.ts
@@ -51,6 +51,167 @@ describe("src/openapi/openapi", () => {
         }
       });
 
+      const errorBody = openApi.bodySchema(
+        "Failed Operation",
+        Types.Object({
+          description: "Error object",
+          properties: {
+            id: Types.Uuid({ description: "Customer ID" }),
+            name: Types.String({
+              description: "Customer name",
+              maxLength: 100,
+              required: true
+            })
+          }
+        })
+      );
+
+      openApi.addPath(
+        "/test",
+        {
+          post: {
+            description: "Test endpoint",
+            operationId: "postid",
+            requestSchema: { body },
+            responses: {
+              200: textPlain("Successful operation.")
+            },
+            summary: "Server Test",
+            tags: ["Internals"]
+          },
+          patch: {
+            description: "Test endpoint",
+            operationId: "patchid",
+            requestSchema: { body },
+            responses: {
+              200: textPlain("Successful operation."),
+              400: errorBody
+            },
+            summary: "Server Test",
+            tags: ["Internals"]
+          }
+        },
+        true
+      );
+      expect(openApi.generateJson()).toMatchSnapshot();
+    });
+
+    test.only("object simple, with named response model", async () => {
+      const body = Types.Object({
+        properties: {
+          float: Types.Number(),
+          integer: Types.Integer(),
+          string: Types.String(),
+          binary: Types.Binary(),
+          byte: Types.Byte(),
+          boolean: Types.Boolean(),
+          date: Types.Date(),
+          dateTime: Types.DateTime(),
+          stringarray: Types.Array({ arrayType: Types.String() }),
+          base64array: Types.Array({ arrayType: Types.Byte() }),
+          internalobject: Types.Object({
+            properties: {
+              uuid: Types.Uuid(),
+              boolean: Types.Boolean()
+            }
+          }),
+          objectArray: Types.Array({
+            arrayType: Types.Object({
+              properties: {
+                obj: Types.Object({
+                  properties: {
+                    uuid: Types.Uuid(),
+                    boolean: Types.Boolean()
+                  }
+                })
+              }
+            })
+          })
+        }
+      });
+
+      const errorBody = openApi.bodySchema(
+        "Failed Operation",
+        Types.Object({
+          description: "Error object",
+          properties: {
+            id: Types.Uuid({ description: "Customer ID" }),
+            name: Types.String({
+              description: "Customer name",
+              maxLength: 100,
+              required: true
+            })
+          },
+          modelName: "ErrorResponse"
+        })
+      );
+
+      openApi.addPath(
+        "/test",
+        {
+          post: {
+            description: "Test endpoint",
+            operationId: "postid",
+            requestSchema: { body },
+            responses: {
+              200: textPlain("Successful operation.")
+            },
+            summary: "Server Test",
+            tags: ["Internals"]
+          },
+          patch: {
+            description: "Test endpoint",
+            operationId: "patchid",
+            requestSchema: { body },
+            responses: {
+              200: textPlain("Successful operation."),
+              400: errorBody
+            },
+            summary: "Server Test",
+            tags: ["Internals"]
+          }
+        },
+        true
+      );
+      expect(openApi.generateJson()).toMatchSnapshot();
+    });
+
+    test("object simple, with models", async () => {
+      const body = Types.Object({
+        properties: {
+          float: Types.Number(),
+          integer: Types.Integer(),
+          string: Types.String(),
+          binary: Types.Binary(),
+          byte: Types.Byte(),
+          boolean: Types.Boolean(),
+          date: Types.Date(),
+          dateTime: Types.DateTime(),
+          stringarray: Types.Array({ arrayType: Types.String() }),
+          base64array: Types.Array({ arrayType: Types.Byte() }),
+          internalobject: Types.Object({
+            properties: {
+              uuid: Types.Uuid(),
+              boolean: Types.Boolean()
+            }
+          }),
+          objectArray: Types.Array({
+            arrayType: Types.Object({
+              properties: {
+                obj: Types.Object({
+                  properties: {
+                    uuid: Types.Uuid(),
+                    boolean: Types.Boolean()
+                  }
+                })
+              },
+              modelName: "UuidObject"
+            }),
+            modelName: "UuidList"
+          })
+        }
+      });
+
       openApi.addPath(
         "/test",
         {
